+++ NEWS.md
@@ -8,6 +8,9 @@ Date: 2019.02.17 (current development)
 
 Changes:
 *   bestModel function noew accepts a formula as a predictors object
+*   plotPercentile now accepts descending ranking
+*   rawTable and normTable adapted for descending values
+*   rangeCheck prints additional information